+++ sdk/search/Azure.Search.Documents/tests/Samples/HelloWorld.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using Azure.Core.TestFramework;
 #region Snippet:Azure_Search_Tests_Samples_Namespaces
 using Azure.Search.Documents;
-using Azure.Search.Documents.Models;
+using Azure.Search.Documents.Indexes.Models;
 #endregion Snippet:Azure_Search_Tests_Samples_Namespaces
 using NUnit.Framework;